% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{allSampleLineCharts}
\alias{allSampleLineCharts}
\title{Draw Line Plot For All Samples}
\usage{
allSampleLineCharts(
  clusterCI,
  fillColor = NULL,
  saveDir = NULL,
  saveWidth = 8,
  saveHeight = 7,
  pointSize = 3,
  pointShape = 16,
  lineSize = 1,
  lineShape = 1,
  pointAlpha = 1,
  lineAlpha = 1,
  errorbarSize = 1,
  errorbarType = 1,
  yLimit = NULL,
  fontSize = 4,
  fontYMove = -0.02,
  legendPosition = "right",
  xAxisFontSize = 14,
  yAxisFontSize = 14,
  xAxisScaleSize = 12,
  yAxisScaleSize = 12,
  legendTitleSize = 14,
  legendTextSize = 14
)
}
\arguments{
\item{clusterCI}{A data.frame of the subclone clustering results is
consistent with the 'ci' in the list returned by runTumorCluster() or
postProcess().}

\item{fillColor}{The fill color of the line plot, different colors for
different clusters, and the number of colors must not be less than the
number of clusters. If not specified, built-in colors are used.}

\item{saveDir}{A value for a local directory to which the image in pdf format
will be saved, or if not specified, the image will be output to the screen.}

\item{saveWidth}{The width of the saved pdf image, valid only if 'saveDir'
is not NULL, default is 8.}

\item{saveHeight}{The length of the saved pdf image, valid only if 'saveDir'
is not NULL, default is 7.}

\item{pointSize}{The size of the point, default is 3.}

\item{pointShape}{The shape of the point, default is 16.}

\item{lineSize}{The size of the line, default is 1.}

\item{lineShape}{The shape of the line, default is 1}

\item{pointAlpha}{The transparency of the color of the point, defaults to 1.}

\item{lineAlpha}{The transparency of the color of the line, defaults to 1.}

\item{errorbarSize}{The size of the error bars of CCF, default is 1.}

\item{errorbarType}{The type of the error bars of CCF, default is 1.}

\item{yLimit}{A vector specifying the range of the y-axis, i.e., the display
range of the CCF, such as c(0, 1). If not specified, it will be automatically
adjusted based on the data.}

\item{fontSize}{The font size of the text in the confidence interval, default is 4.}

\item{fontYMove}{The distance of the text of the confidence interval compared
to the y-axis of the corresponding point, default is -0.02.}

\item{legendPosition}{The position of legends ("none", "left", "right",
"bottom", "top", "inside"), default is "right".}

\item{xAxisFontSize}{The x-axis title font size, default is 14.}

\item{yAxisFontSize}{The y-axis title font size, default is 14.}

\item{xAxisScaleSize}{The x-axis ruler font size, default is 12.}

\item{yAxisScaleSize}{The y-axis ruler font size, default is 12.}

\item{legendTitleSize}{The legend title font size, default is 14.}

\item{legendTextSize}{The legend text font size, default is 14.}
}
\value{
Save to a pdf file in a local directory or output an image in
'Plots' of RStudio.
}
\description{
Draw Line Plot For All Samples
}
