% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{allTreeGraph}
\alias{allTreeGraph}
\title{Draw All Clone Evolution Trees}
\usage{
allTreeGraph(
  treeList,
  saveDir = NULL,
  saveWidth = 8,
  saveHeight = 7,
  margin = c(0, 0, 0, 0),
  asp = c(1, 1),
  vertexColor = NULL,
  vertexAlpha = 0.8,
  vertexFrameColor = "white",
  vertexShape = "circle",
  vertexSize = 45,
  vertexLabelCex = 2,
  vertexLabelColor = "black",
  edgeColor = "black",
  egdeWidth = 3,
  edgeArrowSize = 0.8,
  edgeArrowWidth = 0.8,
  edgeLty = 1,
  edgeCurved = 0
)
}
\arguments{
\item{treeList}{A list of all clone evolutionary trees returned by enumerateAllTree().}

\item{saveDir}{A value for a local directory to which the image in pdf format
will be saved, or if not specified, the image will be output to the screen.}

\item{saveWidth}{The width of the saved pdf image, valid only if 'saveDir'
is not NULL, default is 8.}

\item{saveHeight}{The length of the saved pdf image, valid only if 'saveDir'
is not NULL, default is 7.}

\item{margin}{The distance vector of the image from the boundary, the
default is c(0, 0, 0, 0).}

\item{asp}{The vector of the aspect ratio of the image (y/x), default is c(1, 1).}

\item{vertexColor}{The fill color of the node, different colors for
different clusters, and the number of colors must not be less than the
number of clusters. If not specified, built-in colors are used.}

\item{vertexAlpha}{The transparency of the color of the node, default is 0.8.}

\item{vertexFrameColor}{The color of the node's border, default is "white".}

\item{vertexShape}{The shape of the node("none","circle","square","csquare",
"rectangle","crectangle","vrectangle","pie","raster","sphere"), default is "circle".}

\item{vertexSize}{The size of the node, default is 45.}

\item{vertexLabelCex}{The font size of the node label, default is 2}

\item{vertexLabelColor}{The font color of the node label, default is "black".}

\item{edgeColor}{The color of the edge, default is "black".}

\item{egdeWidth}{The width of the edge, default is 3.}

\item{edgeArrowSize}{The size of the side arrow, default is 0.8.}

\item{edgeArrowWidth}{The width of the side arrow, default is 0.8.}

\item{edgeLty}{The type of the line of the edge, default is 1.}

\item{edgeCurved}{The curvature of the edge, range 0-1, default 0.}
}
\value{
Save to a pdf file in a local directory or output an image in
'Plots' of RStudio.
}
\description{
Draw All Clone Evolution Trees
}
