% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{calcCI}
\alias{calcCI}
\title{Calculate The CI of The CCF of The Cluster}
\usage{
calcCI(ccfMat, bootstrapIter = 2000, bootstrapConf = 0.95)
}
\arguments{
\item{ccfMat}{ccf matrix.}

\item{bootstrapIter}{bootstrapIter.}

\item{bootstrapConf}{bootstrapConf.}
}
\description{
Calculate The CI of The CCF of The Cluster
}
