% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{calcCloneProp}
\alias{calcCloneProp}
\title{Calculation of Subclone Ratio}
\usage{
calcCloneProp(clusterMat, evolution)
}
\arguments{
\item{clusterMat}{A data.frame of the subclone clustering results is
consistent with the 'mat' in the list returned by postProcess().}

\item{evolution}{An evolutionary tree data.frame returned by
generateDefaultTree() or enumerateAllTree().}
}
\value{
Returns a data.frame in which the proportion of subclones in each
cluster for each sample has been calculated.
}
\description{
Calculate the subclone proportion of each sample based on any
evolutionary tree.
}
