% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{checkCloneCCF}
\alias{checkCloneCCF}
\title{Checking for Clone Cluster Problems}
\usage{
checkCloneCCF(clusterResult)
}
\arguments{
\item{clusterResult}{clusterResult.}
}
\description{
Check for clonal cluster problems sample by sample, merging
clusters where the ccf of subclonal clusters is greater than the ccf of
clonal clusters
}
