% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{convert2SciCloneInput}
\alias{convert2SciCloneInput}
\title{Convert to SciClone Input Format}
\usage{
convert2SciCloneInput(
  inputList,
  calcCCFMethod = "distance",
  maxCCFValue = 1.2,
  LimitCCF = TRUE
)
}
\arguments{
\item{inputList}{Same as 'inputList' parameter of runTumorCluster()}

\item{calcCCFMethod}{calcCCFMethod.}

\item{maxCCFValue}{maxCCFValue}
}
\value{
Returns a list containing the snv list and anno list needed for
sciclone clustering as well as the original input list with calculated CCF
and mutation multiplicity.
}
\description{
Convert to SciClone Input Format
}
