% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{doubleSampleScatter}
\alias{doubleSampleScatter}
\title{Draw Scatter Plot of Pairwise Samples}
\usage{
doubleSampleScatter(
  clusterMat,
  fillColor = NULL,
  saveDir = NULL,
  saveWidth = 8,
  saveHeight = 7,
  pointSize = 2.5,
  pointShape = 17,
  alpha = 0.8,
  legendPosition = "right",
  xLimit = NULL,
  yLimit = NULL,
  titleSize = 15,
  xAxisFontSize = 14,
  yAxisFontSize = 14,
  xAxisScaleSize = 12,
  yAxisScaleSize = 12,
  legendTitleSize = 14,
  legendTextSize = 14
)
}
\arguments{
\item{clusterMat}{A data.frame of the subclone clustering results is
consistent with the 'mat' in the list returned by runTumorCluster() or
postProcess().}

\item{fillColor}{The fill color of the Scatter plot, different colors for
different clusters, and the number of colors must not be less than the
number of clusters. If not specified, built-in colors are used.}

\item{saveDir}{A value for a local directory to which the image in pdf format
will be saved, or if not specified, the image will be output to the screen.}

\item{saveWidth}{The width of the saved pdf image, valid only if 'saveDir'
is not NULL, default is 8.}

\item{saveHeight}{The length of the saved pdf image, valid only if 'saveDir'
is not NULL, default is 7.}

\item{pointSize}{The size of the point, default is 2.5.}

\item{pointShape}{The shape of the point, default is 17.}

\item{alpha}{The transparency of the color of the point, default is 0.8.}

\item{legendPosition}{The position of legends ("none", "left", "right",
"bottom", "top", "inside"), default is "right".}

\item{xLimit}{A vector specifying the range of the x-axis, i.e., the display
range of the CCF, such as c(0, 1). If not specified, it will be automatically
adjusted based on the data.}

\item{yLimit}{A vector specifying the range of the y-axis, i.e., the display
range of the CCF, such as c(0, 1). If not specified, it will be automatically
adjusted based on the data.}

\item{titleSize}{The title text size, default is 15.}

\item{xAxisFontSize}{The x-axis title font size, default is 14.}

\item{yAxisFontSize}{The y-axis title font size, default is 14.}

\item{xAxisScaleSize}{The x-axis ruler font size, default is 12.}

\item{yAxisScaleSize}{The y-axis ruler font size, default is 12.}

\item{legendTitleSize}{The legend title font size, default is 14.}

\item{legendTextSize}{The legend text font size, default is 14.}
}
\value{
Save to a pdf file in a local directory or output an image in
'Plots' of RStudio.
}
\description{
Draw Scatter Plot of Pairwise Samples
}
