% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{enumerateAllTree}
\alias{enumerateAllTree}
\title{Enumerate and Sort All Spanning Trees}
\usage{
enumerateAllTree(
  adj_matrix,
  error_buf = 0.1,
  max_tree_num = 20,
  remove_cluster = FALSE,
  inferTreeInput = NULL
)
}
\arguments{
\item{adj_matrix}{A data.frame of adjacency matrix generated by
GenerateAdjMatrix(),representing the containment relationship between clusters.}

\item{error_buf}{Size of the error buffer,default is 0.1. Represent the CCF
of the parent cluster as the CCF of the parent cluster multiplied by
'error_buf' to account for error noise.}

\item{max_tree_num}{The maximum number of cloned evolutionary trees to
enumerate if there are multiple different evolutionary trees, default is 20.}

\item{remove_cluster}{A Boolean value indicating whether it is permissible to
delete the cluster with the lowest number of SNVs and then re-run the
enumeration tree process when the current cluster node is unable to generate
an evolutionary tree.}

\item{inferTreeInput}{A data.frame of 'inferTree' that is consistent with
the input data of GenerateAdjMatrix(). Valid only if 'remove_cluster' is TRUE.}
}
\value{
Returns a list that includes a sorted list of evolutionary trees and
a sorted weights data.frame, where each data.frame in the evolutionary tree
list represents an evolutionary tree.
}
\description{
Based on the adjacency matrix, enumerate all the cloned
evolutionary trees that satisfy the sum condition and the crossover rule,
and sort them according to the probability of occurrence of the total edges.
}
