% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{extractFromSC}
\alias{extractFromSC}
\title{Extract the Required Columns From the SC}
\usage{
extractFromSC(sc)
}
\arguments{
\item{sc}{cluster result of sciclone.}
}
\description{
Extract the Required Columns From the SC
}
