% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{findMaxSumChild}
\alias{findMaxSumChild}
\title{Finding the Maximum Sum Sub Node}
\usage{
findMaxSumChild(adj_matrix, parent_index, error_buf)
}
\arguments{
\item{adj_matrix}{The adjacency matrix generated by GenerateAdjMatrix().}

\item{parent_index}{The ID of the parent node cluster.}

\item{error_buf}{Size of the error buffer.}
}
\value{
Returns a vector of child node IDs, where the child nodes are
children of the largest sum of ccf's of that parent node.
}
\description{
Find the child node that does not exceed the maximized ccf sum
of the parent node.
}
