% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{GenerateAdjMatrix}
\alias{GenerateAdjMatrix}
\title{Generate Adjacency Matrix}
\usage{
GenerateAdjMatrix(inferTreeInput)
}
\arguments{
\item{inferTreeInput}{A data.frame, which is the 'inferTree' data.frame from
the list returned by the postProcess() function, containing the mutation
information and CCF for each sample.}
}
\value{
Returns a list containing a data.frame for the adjacency matrix and
a data.frame for the CCF of clusters.
}
\description{
Generate an adjacency matrix representing the containment
relationships between clusters using the inferred evolutionary tree
obtained from the post-clustering processing steps.
}
