% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{generateDefaultTree}
\alias{generateDefaultTree}
\title{Generating the Default Tree}
\usage{
generateDefaultTree(adj_matrix, error_buf = 0.1)
}
\arguments{
\item{adj_matrix}{A data.frame of adjacency matrix generated by
GenerateAdjMatrix(), representing the containment relationship between clusters.}

\item{error_buf}{Size of the error buffer, default is 0.1.Represent the CCF
of the parent cluster as the CCF of the parent cluster multiplied by
'error_buf' to account for error noise.}
}
\value{
Returns a data.frame where the first column represents the parent
cluster and the second column represents the child clusters.
}
\description{
Quickly find a clonal evolutionary tree from the adjacency
matrix that meets the sum conditions and crossing rules based on a greedy
strategy.
}
