% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{generateSampleTree}
\alias{generateSampleTree}
\title{Inferring Sample Evolutionary Trees}
\usage{
generateSampleTree(sampleGroup)
}
\arguments{
\item{sampleGroup}{A list of grouping cases returned by runSampleGroup().}
}
\value{
Returns an igraph tree object.
}
\description{
Infer the common ancestor and evolutionary relationship between
samples based on the grouping returned by runSampleGroup().
}
