% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{install_sciClone}
\alias{install_sciClone}
\title{Install SciClone}
\usage{
install_sciClone()
}
\description{
Automatically install sciclone. If the installation fails,
please refer to the sciclone installation guide(https://github.com/genome/sciclone).
}
