% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{isSumRule}
\alias{isSumRule}
\title{Check That The Sum Condition is Satisfied}
\usage{
isSumRule(edge_df, cluster_ccf, error_buf)
}
\arguments{
\item{edge_df}{Cloned evolutionary tree data.frame with parent node in the
first column and child nodes in the second column.}

\item{cluster_ccf}{CCF for each cluster.}

\item{error_buf}{Size of the error buffer.}
}
\value{
Returns a boolean value, if TRUE means the sum condition is satisfied.
}
\description{
Check That The Sum Condition is Satisfied
}
