% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{mergeCluster}
\alias{mergeCluster}
\title{Merge Cluster}
\usage{
mergeCluster(clusterResult, pvalue = 0.05)
}
\arguments{
\item{clusterResult}{clusterResult.}

\item{pvalue}{pvalue.}
}
\description{
Merge Cluster
}
