% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotSubcloneProp}
\alias{plotSubcloneProp}
\title{Draw Subclone proportion histogram}
\usage{
plotSubcloneProp(
  prop,
  position = "fill",
  fillColor = NULL,
  saveDir = NULL,
  saveWidth = 8,
  saveHeight = 7,
  alpha = 0.8,
  legendPosition = "right",
  title = "Clone Proportion",
  titleSize = 15,
  xAxisFontSize = 14,
  yAxisFontSize = 14,
  xAxisScaleSize = 12,
  legendTitleSize = 14,
  legendTextSize = 14
)
}
\arguments{
\item{prop}{A data.frame of the proportion of each cluster for each sample
returned by calcCloneProp().}

\item{position}{Position parameters of the histogram ("fill", "stack",
"dodge"), default is "fill".}

\item{fillColor}{The fill color of the histogram, different colors for
different clusters, and the number of colors must not be less than the
number of clusters.}

\item{saveDir}{A value for a local directory to which the image in pdf format
will be saved, or if not specified, the image will be output to the screen.}

\item{saveWidth}{The width of the saved pdf image,,valid only if 'saveDir'
is not NULL,default is 8.}

\item{saveHeight}{The length of the saved pdf image,valid only if 'saveDir'
is not NULL,default is 7.}

\item{alpha}{The transparency of the color, default is 0.8.}

\item{legendPosition}{The position of legends ("none", "left",
"right", "bottom", "top", "inside"),default is "right".}

\item{title}{The title of the image, default is "Clone Proportion".}

\item{titleSize}{The title text size, default is 15.}

\item{xAxisFontSize}{The x-axis title font size, default is 14.}

\item{yAxisFontSize}{The y-axis title font size, default is 14.}

\item{xAxisScaleSize}{The x-axis ruler font size, default is 12.}

\item{legendTitleSize}{The legend title font size, default is 14.}

\item{legendTextSize}{The legend text font size, default is 14.}
}
\value{
Save to a pdf file in a local directory or output an image in
'Plots' of RStudio.
}
\description{
Draw Subclone proportion histogram
}
