% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{postProcess}
\alias{postProcess}
\title{Clustering Post-Processing}
\usage{
postProcess(
  clusterResult,
  removeOutliers = TRUE,
  minSnvNum = 5,
  mergeUndiffClust = TRUE,
  mergePvalue = 0.05,
  bootstrapIter = 2000,
  bootstrapConf = 0.95
)
}
\arguments{
\item{clusterResult}{A list returned by runTumorCluster().}

\item{removeOutliers}{Boolean value, if TRUE then outliers are removed based.
on quartile distance, if FALSE then they are not removed.}

\item{minSnvNum}{Minimum number of SNVs in the cluster, clusters with SNVs
less than this value will be deleted, default is 5, if you don't want to
delete the cluster you can set it to 0.}

\item{mergeUndiffClust}{Boolean value, if TRUE then clusters that are not
statistically different are merged according to the wilcox test,
if FALSE then they are not merged.}

\item{mergePvalue}{The p-value for the wilcox test, valid only if
mergeUndiffClust=TRUE. Default is 0.05.}

\item{bootstrapIter}{The number of bootstrap iterations, default is 2000.}

\item{bootstrapConf}{Confidence interval for bootstrap, default is 0.95.}
}
\value{
Returns a list containing one more data.frame, used to infer the
evolutionary tree, than the input list (clusterResult). The clustering
results and confidence intervals in the list have all been processed.
}
\description{
Post-processing of the clustering results included
removing outlier points; removing points with too few mutations in the cluster;
merging clusters that were not statistically different;
and merging irrational subclone clusters into clonal clusters.
}
