% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{RunPipeline}
\alias{RunPipeline}
\title{Run the Pipeline of RETCHER}
\usage{
RunPipeline(
  inputList,
  saveDir = NULL,
  calcCCFMethod = "distance",
  maxCCFValue = 1.2,
  sampleNames = NULL,
  minimumDepth = 20,
  maximumClusters = 10,
  useSexChrs = TRUE,
  bootstrapIter = 2000,
  bootstrapConf = 0.95,
  usePercentage = TRUE,
  LimitCCF = TRUE,
  clusterMinSnvNum = 5,
  removeOutliers = TRUE,
  mergeUndiffClust = TRUE,
  mergePvalue = 0.05,
  error_buf = 0.1,
  enumAllTree = TRUE,
  maxTreeNum = 20,
  allowRemoveCluster = TRUE,
  minPresenceCCF = 0.05
)
}
\arguments{
\item{inputList}{A list of data.frames, where each data.frame represents a
sample, and each sample has 8 columns of required data.}

\item{saveDir}{Save directory for run results, all files will automatically
go to this directory, if the directory doesn't exist then it will be created
automatically, if the directory is NULL, then it will only output the result
image to RStudio.}

\item{calcCCFMethod}{Methods for calculating Cancer Cell Fraction (CCF) are
supported ("distance", "threshold"). The difference between them lies in the
method of calculating mutation multiplicity, with the default being "distance".}

\item{maxCCFValue}{The maximum CCF value of SNV. SNVs larger than this value
will be filtered. The default is 1.2.}

\item{sampleNames}{A vector of sample names, the length of which should be
the same as the number of samples, or the names of the elements of the
'inputList' if not specified.}

\item{minimumDepth}{The minimum sequencing depth. SNV with a depth lower
than this will be filtered out. The default is 20.}

\item{maximumClusters}{The maximum number of clusters for subclone clustering.
the default is 10.}

\item{useSexChrs}{Boolean parameter to specify whether (TRUE) or not (FALSE)
mutations on the sex chromosomes are used in the clustering process. The
default is TRUE.}

\item{bootstrapIter}{The number of bootstrap iterations. the default is 2000.}

\item{bootstrapConf}{The confidence interval for bootstrap. the default is 0.95.}

\item{usePercentage}{Whether to multiply CCF by 100 to express percentage, default is TRUE}

\item{LimitCCF}{Whether to limit CCF to 0 to 1, default is TRUE.}

\item{clusterMinSnvNum}{The minimum number of SNV in a subclone cluster.
clusters with fewer SNV than this value will be removed. If you do not want
to remove clusters, you can set it to 0. The default is 5.}

\item{removeOutliers}{Boolean parameter to specify whether (TRUE) or not
(FALSE) to remove outliers based on the interquartile range. The default is TRUE.}

\item{mergeUndiffClust}{Boolean parameter to specify whether (TRUE) or not
(FALSE) to merge clusters without statistical differences based on the
Wilcox test. The default is TRUE.}

\item{mergePvalue}{The p-value threshold for the Wilcox test, above which
clusters are considered to have no statistical difference. This parameter is
effective only when the 'mergeUndiffClust' parameter is set to TRUE.
The default is 0.05.}

\item{error_buf}{The error buffer used to determine if clusters violate the
sum conditions between them. The default value is 0.1.}

\item{enumAllTree}{Boolean parameter to specify whether (TRUE) or not (FALSE)
to enumerate all possible clone evolutionary trees. The default is TRUE.}

\item{maxTreeNum}{The maximum number of clone evolutionary trees to enumerate
if there are multiple different trees. The default is 20.}

\item{allowRemoveCluster}{Boolean parameter to specify whether (TRUE) or not
(FALSE) to automatically remove clusters with the fewest SNV and re-run the
tree enumeration process if it is not possible to generate a clone
evolutionary tree. The default is TRUE.}

\item{minPresenceCCF}{The minimum CCF at which an SNV is considered to exist
in the sample; SNVs with a CCF below this value are considered not to exist
in the sample. The default is 0.05.}
}
\value{
Returns different results according to different situations.Typically
returns a list containing subclone cluster results; cluster post-processing
results; adjacency matrix; default clonal evolution tree; all possible
evolution trees that have been sorted; subclone proportions calculated with
the default evolution tree; sample group; sample tree. The results of each
section can be re-run using the corresponding function.
}
\description{
RETCHER is used for reconstructing tumor clonal heterogeneity
and evolutionary relationships. This function integrates the various steps
of RETCHER into a pipeline for ease of use.
}
\details{
In pipeline, all parameters except the declared parameters are
default parameters, especially the drawing parameters, and the user can
still run the drawing function separately after executing pipeline.

If the parameter saveDir is not NULL, RETCHER creates three folders
(cluster, tree, plot) in the saveDir directory, corresponding to the
results of the clustering, inferred tree, and plotting, respectively.

The pipeline uses the default evolutionary tree when calculating subclone
proportions.
}
