% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{runTumorCluster}
\alias{runTumorCluster}
\title{Run Subclone Clustering}
\usage{
runTumorCluster(
  inputList,
  calcCCFMethod = "distance",
  maxCCFValue = 1.2,
  sampleNames = NULL,
  minimumDepth = 20,
  maximumClusters = 10,
  useSexChrs = TRUE,
  bootstrapIter = 2000,
  bootstrapConf = 0.95
)
}
\arguments{
\item{inputList}{A list of data.frames, where each data.frame represents a
sample, and each sample has 8 columns of required data.}

\item{calcCCFMethod}{Methods for calculating Cancer Cell Fraction (CCF) are
supported ("distance", "threshold"). The difference between them lies in the
method of calculating mutation multiplicity, with the default being "distance".}

\item{maxCCFValue}{The maximum CCF value of SNV. SNV larger than this value
will be filtered. The default is 1.2.}

\item{sampleNames}{A vector of sample names, the length of which should be the
same as the number of samples, or the names of the elements of the "inputList"
if not specified.}

\item{minimumDepth}{Minimum sequencing depth below which SNVs will be filtered,
default is 20.}

\item{maximumClusters}{Maximum number of clusters, default is 10.}

\item{useSexChrs}{Boolean parameter to specify whether (TRUE) or not (FALSE)
mutations on the sex chromosomes are used in the clustering process. default is TRUE.}

\item{bootstrapIter}{The number of bootstrap iterations, default is 2000.}

\item{bootstrapConf}{Confidence interval for bootstrap, default is 0.95.}
}
\value{
Returns a list containing a data.frame of the clustering results,
a data.frame of the confidence intervals for each subclone cluster, a list
of CCF and mutation multiplicities for each SNV for each sample, and a
SciClone clustering results object.
}
\description{
Run subclone clustering from a list of input data.
}
