% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{sampleTreeGraph}
\alias{sampleTreeGraph}
\title{Draw Sample Evolutionary Tree}
\usage{
sampleTreeGraph(
  gGraph,
  title = "Sample Tree",
  titleSize = 20,
  saveDir = NULL,
  saveWidth = 8,
  saveHeight = 7,
  layoutType = "tree",
  margin = c(0, 0, 0, 0),
  asp = c(1, 1),
  vertexColor = NULL,
  vertexAlpha = 0.8,
  vertexSize = NULL,
  vertexFrameColor = "white",
  vertexShape = "circle",
  vertexLabelCex = 1.1,
  vertexLabelColor = "black",
  edgeColor = "black",
  egdeWidth = 3,
  edgeArrowSize = 0.8,
  edgeArrowWidth = 0.8,
  edgeLty = 1,
  edgeCurved = 0
)
}
\arguments{
\item{gGraph}{The tree object of igraph returned by generateSampleTree().}

\item{title}{The title of the image, default is "Sample Tree".}

\item{titleSize}{The title text size, default is 20.}

\item{saveDir}{A value for a local directory to which the image in pdf format
will be saved, or if not specified, the image will be output to the screen.}

\item{saveWidth}{The width of the saved pdf image, valid only if 'saveDir'
is not NULL, default is 8.}

\item{saveHeight}{The length of the saved pdf image, valid only if 'saveDir'
is not NULL, default is 7.}

\item{layoutType}{Layout of igarph("tree","star","circle","sugiyama","nicely",
"randomly"), default is "tree". In general, "sugiyama" is good too.}

\item{margin}{The distance vector of the image from the boundary, the
default is c(0, 0, 0, 0).}

\item{asp}{The vector of the aspect ratio of the image (y/x), default is c(1, 1).}

\item{vertexColor}{The fill color of the node, different colors for
different group, and the number of colors must not be less than the
number of group. If not specified, built-in colors are used.}

\item{vertexAlpha}{The transparency of the color of the node, default is 0.8.}

\item{vertexSize}{The size of the node, if not set, it will change according
to the number of mutations.}

\item{vertexFrameColor}{The color of the node's border, default is "white".}

\item{vertexShape}{The shape of the node("none","circle","square","csquare",
"rectangle","crectangle","vrectangle","pie","raster","sphere"), default is "circle".}

\item{vertexLabelCex}{The font size of the node label, default is 1.1}

\item{vertexLabelColor}{The font color of the node label, default is "black".}

\item{edgeColor}{The color of the edge, default is "black".}

\item{egdeWidth}{The width of the edge, default is 3.}

\item{edgeArrowSize}{The size of the side arrow, default is 0.8.}

\item{edgeArrowWidth}{The width of the side arrow, default is 0.8.}

\item{edgeLty}{The type of the line of the edge, default is 1.}

\item{edgeCurved}{The curvature of the edge, range 0-1, default 0.}
}
\value{
Save to a pdf file in a local directory or output an image in
'Plots' of RStudio.
}
\description{
Draw Sample Evolutionary Tree
}
