% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{saveAdjmat}
\alias{saveAdjmat}
\title{Save Adjacency Matrix}
\usage{
saveAdjmat(adjmat, saveDir)
}
\arguments{
\item{adjmat}{A list returned by GenerateAdjMatrix().}

\item{saveDir}{Save directory, automatically save all files to this directory,
and automatically create it if it does not exist.}
}
\description{
Save the adjacency matrix along with the average CCF for each
cluster.
}
