% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{saveAllTree}
\alias{saveAllTree}
\title{Save All Spanning Trees}
\usage{
saveAllTree(sortedAllTree, saveDir)
}
\arguments{
\item{sortedAllTree}{A list returned by enumerateAllTree().}

\item{saveDir}{Save directory, automatically save all files to this directory,
and automatically create it if it does not exist.}
}
\description{
Save all spanning tree and sort weight data frames.
}
