% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{saveCloneProp}
\alias{saveCloneProp}
\title{Save Subclone Ratio}
\usage{
saveCloneProp(prop, saveDir)
}
\arguments{
\item{prop}{A data.frame returned by calcCloneProp().}

\item{saveDir}{Save directory, automatically save all files to this directory,
and automatically create it if it does not exist.}
}
\description{
Save Subclone Ratio
}
