% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferTree.R
\name{saveDefaultTree}
\alias{saveDefaultTree}
\title{Save Default Tree}
\usage{
saveDefaultTree(defaultTree, saveDir)
}
\arguments{
\item{defaultTree}{A default tree data.frame returned by generateDefaultTree().}

\item{saveDir}{Save directory, automatically save all files to this directory,
and automatically create it if it does not exist.}
}
\description{
Save Default Tree
}
