% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{saveSampleGroupRes}
\alias{saveSampleGroupRes}
\title{Saving the Results of Samples Grouping}
\usage{
saveSampleGroupRes(sampleGroup, saveDir)
}
\arguments{
\item{sampleGroup}{A list returned by runSampleGroup().}

\item{saveDir}{Save directory, automatically save all files to this directory,
and automatically create it if it does not exist.}
}
\description{
Save sample grouping data.frame and average CCF per group data.frame.
}
