% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{saveTumorClusterRes}
\alias{saveTumorClusterRes}
\title{Saving the Results of Tumor Subclone Clustering}
\usage{
saveTumorClusterRes(clusterResult, saveDir)
}
\arguments{
\item{clusterResult}{A list returned by runTumorCluster() or postProcess().}

\item{saveDir}{Save directory, automatically save all files to this directory,
and automatically create it if it does not exist.}
}
\description{
Save subclone clustering results, confidence intervals for
subclone clusters, mutation multiplicity and CCF values for each sample.
}
