% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{extractFromSC}
\alias{extractFromSC}
\title{Extract the Required Columns From the SC}
\usage{
extractFromSC(sc, usePercentage = TRUE)
}
\arguments{
\item{sc}{cluster result of sciclone.}

\item{usePercentage}{Whether to multiply CCF by 100 to express percentage, default is TRUE}
}
\description{
Extract the Required Columns From the SC
}
